Use Instantiation Mode - Put ColorIncreaser into it's own file - Coverage to 100% Fixes #1
