+++ index.html
@@ -5,6 +5,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://cdn.jsdelivr.net/npm/p5@0.4.20/lib/p5.js"></script>
+
+    <script src="ColorIncreaser.js"></script>
     <script src="sketch.js"></script>>
     <style>
         body {
