+++ sketch.js
@@ -7,52 +7,25 @@
 // fillColor will be the color of the rectangle.
 // colorIncreaser will become an instance of our ColorIncreaser class.
 
-let colorValueIncrease = 1;
-let fillColor;
-let colorIncreaser;
+let sketch = function (p) {
+    let colorValueIncrease = 1;
+    let fillColor;
+    let colorIncreaser;
 
-function setup() {
-    createCanvas(500, 500);
-    background(0);
-    fillColor = color(0, 0, 0, 255);
-    noStroke();
+    p.setup = function () {
+        p.createCanvas(500, 500);
+        p.background(0);
+        fillColor = p.color(0, 0, 0, 255);
+        p.noStroke();
 
-    colorIncreaser = new ColorIncreaser(colorValueIncrease, fillColor);
-}
-
-function draw() {
-    fill(colorIncreaser.fillColor);
-    rect(0, 0, 500, 500);
-    colorIncreaser.increaseFillColor();
-}
-
-class ColorIncreaser {
-    constructor(colorValueIncrease, fillColor) {
-        // Stores a value and a color and allows you to increase the color
-        // by that value.
-        this.colorValueIncrease = colorValueIncrease
-        this.fillColor = fillColor;
+        colorIncreaser = new ColorIncreaser(colorValueIncrease, fillColor);
     }
 
-    increaseFillColor() {
-        // increase the first color channel by one.  If that channel
-        // is now > 255 then increment the next color channel.  Repeat for second
-        // and third channel
-        this.fillColor.levels[0] += this.colorValueIncrease
-        this.numColorsSoFar += 1
-
-        if (this.fillColor.levels[0] > 255) {
-            this.fillColor.levels[0] = 0
-            this.fillColor.levels[1] += this.colorValueIncrease
-        }
-        if (this.fillColor.levels[1] > 255) {
-            this.fillColor.levels[1] = 0
-            this.fillColor.levels[2] += this.colorValueIncrease
-        }
-        if (this.fillColor.levels[2] > 255) {
-            this.fillColor.levels[2] = 0;
-        }
+    p.draw = function () {
+        p.fill(colorIncreaser.fillColor);
+        p.rect(0, 0, 500, 500);
+        colorIncreaser.increaseFillColor();
     }
 }
 
-module.exports = ColorIncreaser;
+let myp5 = new p5(sketch);
